requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
python-multipart
supabase
pypdf
mammoth
requests
fastapi
uvicorn
python-dotenv
